requirements.txt
python-telegram-bot[job-queue]>=20.0
python-dotenv>=1.0.0
httpx>=0.24.0
pytz>=2023.3
websocket-client>=1.0.0
requests>=2.28.0
